Fix correctness bugs identified in README corrections section
Agent-Logs-Url: https://github.com/Rookiechan191/kvstore/sessions/055a1e91-390b-44da-a353-15322c7c5cab

+++ router/hash.go
@@ -10,6 +10,6 @@ func hashKey(key string) uint32 {
 
 func getNode(key string, nodes []string) string {
 	hash := hashKey(key)
-	index := int(hash) % len(nodes)
+	index := int(hash % uint32(len(nodes)))
 	return nodes[index]
 }
